+++ CHANGELOG.md
@@ -4,6 +4,10 @@ NEXT_VERSION_BUMP: PATCH
 ## XX.XX.XX - 20XX-XX-XX
 * [ADDED] Added support for Bizum.
 
+### Financial Connections
+* [FIXED] Preserved `no_eligible_accounts` in `onEvent` error callbacks instead of reporting it as `unexpected_error`.
+* [FIXED] Prevented background authorization-session telemetry failures from triggering `onEvent` error callbacks.
+
 ## 23.19.0 - 2026-09-15
 
 ### All SDKs